localhost:3000に接続すると自動的に/booksにいく　appにまとまっていたのをbook,category,userにわけた　nethod-overrideをインストールしてapiをrestfulに
