+++ front/views/category-form.njk
@@ -16,7 +16,7 @@
     <a href="/books">
       <img src="/public/images/Back-Button1.png" onMouseOver="this.src='/public/images/Back-Button2.png'" onMouseOut="this.src='/public/images/Back-Button1.png'" class="rounded float-end" width="55" height="50"/>
     </a>
-    <a class="btn btn-primary" href="/books">カテゴリを追加する</a>
+    <a class="btn btn-primary" href="/books">本を追加する</a>
   </div>
 </div>
 {% endblock %}
