+++ src/category/category.module.ts
@@ -1,0 +1,16 @@
+import { Module } from "@nestjs/common";
+import { SequelizeModule } from "@nestjs/sequelize";
+import { CategoryService } from "./category.service";
+import { CategoryController } from "./category.controller";
+import { Category } from "../models/category.model";
+
+@Module({
+  imports: [
+    SequelizeModule.forFeature([Category])
+  ],
+  controllers: [CategoryController],
+  providers: [CategoryService],
+  exports: [CategoryService],
+})
+
+export class CategoryModule {}
