+++ src/user/user.module.ts
@@ -1,0 +1,17 @@
+import { Module } from "@nestjs/common";
+import { SequelizeModule } from "@nestjs/sequelize";
+import { UserService } from "./user.service";
+import { UserController } from "./user.controller";
+import { User } from "../models/user.model";
+
+
+@Module({
+  imports: [
+    SequelizeModule.forFeature([User])
+  ],
+  controllers: [UserController],
+  providers: [UserService],
+  exports: [UserService],
+})
+
+export class UserModule {}
